+++ src/Stormancer.Plugins/Analytics/Stormancer.Server.Plugins.Analytics/App.cs
@@ -2,8 +2,15 @@
 
 namespace Stormancer.Server.Plugins.Analytics
 {
+    /// <summary>
+    /// Plugin entry point class.
+    /// </summary>
     public class App
     {
+        /// <summary>
+        /// Plugin entry point.
+        /// </summary>
+        /// <param name="builder"></param>
         public void Run(IAppBuilder builder)
         {
             builder.AddPlugin(new AnalyticsPlugin());