+++ src/Stormancer.Plugins/SpatialOS/Stormancer.Server.Plugins.SpatialOS/Changelog.rst
@@ -9,9 +9,12 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+0.2.2
+-----
+- Use Stormancer.Abstraction.Server 4.2.0.2-pre.
 0.2.1
 -----
-- Use Stormancer.Abstraction.Server 6.1.0
+- Use Stormancer.Abstraction.Server 4.1.0.
 
 0.2.0.1
 ----------